Update EXAMPLES.md
These docs are out of date. The correct way to pass the invitation and organization params to `loginWithRedirect` is to do so directly, rather than as an object within `authorizationParams`.

This is the only way I've successfully navigated a user to the sign-up screen from an Organization invite. The previous example fails to pass the invitation and organization params to the /authorization endpoint.